Happy new year!

+++ src/apps/tools/OswAppStepStats.cpp
@@ -19,10 +19,10 @@ void OswAppStepStats::drawChart() {
     hal->getLocalDate(&dayOfMonth, &weekDay);
 
     for (uint8_t index = 0; index < 7; index++) {
-        uint32_t weekDayStep = hal->environment->getStepsOnDay(index);
-        uint16_t chartStickValue = ((float)(weekDayStep > goalValue ? goalValue : weekDayStep) / goalValue) * chartStickHeight;
+        unsigned int weekDayStep = hal->environment->getStepsOnDay(index);
+        unsigned short chartStickValue = ((float)(weekDayStep > goalValue ? goalValue : weekDayStep) / goalValue) * chartStickHeight;
 
-        uint16_t barColor = OswConfigAllKeys::stepsPerDay.get() <= weekDayStep ? ui->getSuccessColor() : changeColor(ui->getSuccessColor(), 2.85);
+        uint16_t barColor = (unsigned int) OswConfigAllKeys::stepsPerDay.get() <= weekDayStep ? ui->getSuccessColor() : changeColor(ui->getSuccessColor(), 2.85);
 
         chartStickValue = chartStickValue < 2 ? 0 : chartStickValue;
